refactor(actions-helper): move existing function to separate module
