+++ .github/workflows/ci.yml
@@ -13,4 +13,4 @@ jobs:
 
       - uses: moneymeets/moneymeets-composite-actions/lint-python@master
 
-      - run: poetry run python -m pytest --cov --cov-fail-under=95
+      - run: poetry run python -m pytest --cov --cov-fail-under=94
